tests: extract TestResult to separate file
